requirements.txt
#pip install -r requirements.txt
pandas 
scikit-learn 
matplotlib 
google-generativeai 
python-dotenv 
requests